requirements.txt
# 核心依赖
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# 向量数据库
faiss-cpu>=1.7.4
# chromadb>=0.4.0  # 可选

# Web 框架
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
# gradio>=4.0.0  # 可选

# 文档处理
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.0.0  # Pro 版通用对话
duckduckgo-search>=3.9.0  # Pro 版联网搜索（可选）
python-multipart>=0.0.6  # 文件上传支持

# 可选：本地模型支持
# ollama>=0.1.0

# 开发工具
pytest>=7.4.0
black>=23.11.0
mypy>=1.7.0

